pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "A simple OAuth2 authorization server."
authors = [
    { name="Omar Crosby", email="omar.crosby@gmail.com" }
]
dependencies = [
    "bcrypt==4.0.1",
    "email-validator",
    "fastapi",
    "httpx",
    "passlib[bcrypt]",
    "python-jose[cryptography]",
    "python-multipart",
    "requests",
    "sqlmodel",
    "toml",
    "types-toml",
    "uvicorn[standard]",
]

[build-system]
requires = ["flit_core>=3.2,<4"]
build-backend = "flit_core.buildapi"

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "isort",
    "mypy",
    "pylint",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-mock",
    "radon"
]

[project.scripts]
hashpwd = "cli.hash:hash_password"
